+++ R/0_1_globals.R
@@ -35,13 +35,13 @@ utils::globalVariables(c(
   "data_prepared_with_recipe",
 
   # textTrainNPlot
-  "sample_size", "percent", "std",
+  "sample_size", "percent", "std", "ymin", "ymax",
 
   # textPredcitTest
   "stats_on_bootstrap",
 
   # textPredict
-  ":=", "predicted_class",
+  ":=", "predicted_class", "osf_download", "osf_retrieve_file",
 
   # implicit motives
   "resid",